rename files to keep naming conventions
