+++ client/src/app/game/components/table/cards/cards.component.ts
@@ -3,7 +3,7 @@ import { Card, getCardsAssetNumber, Suit } from '../../../../models/card'
 import { DomSanitizer, SafeStyle } from '@angular/platform-browser'
 import { GameService } from '../../../game.service'
 import { AddPlayerMessage, PlayersTurnMessage } from '../../../../models/message'
-import { PlayerAction } from '../../../../models/playerAction'
+import { PlayerAction } from '../../../../models/player-action'
 import { AlertController, ModalController } from '@ionic/angular'
 import { SelectSuitComponent } from '../select-suit/select-suit.component'
 import { SettingsService } from '../../../../settings/settings.service'
