+++ client/src/app/game/helpers/inactivity-detection.ts
