+++ client/src/app/game/helpers/leave-game.guard.ts
